backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
Pillow>=9.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0